+++ pyiron_continuum/damask/factory.py
@@ -10,7 +10,6 @@ with ImportAlarm(
 ) as damask_alarm:
     from damask import GeomGrid, YAML, ConfigMaterial, seeds, Rotation
 import numpy as np
-from pyiron_continuum.toolkit import composition_to_spacegroup
 
 __author__ = "Muhammad Hassani"
 __copyright__ = (
@@ -312,3 +311,18 @@ class Create:
         if isinstance(method, str):
             method = getattr(Rotation, method)
         return method(*args, **kwargs)
+
+
+def get_element_abbreviation(name):
+    import periodictable
+    for element in periodictable.elements:
+        if element.name.lower() == name.lower():
+            return element.symbol
+    raise NameError(name, "does not exist")
+
+
+def composition_to_spacegroup(composition):
+    from ase.build import bulk
+    from ase.spacegroup import get_spacegroup
+    abbreviation = get_element_abbreviation(composition)
+    return get_spacegroup(bulk(abbreviation))
